pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.116.1",
    "mangum>=0.19.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 150


